+++ Control/Joint/Schemes/UT.hs
@@ -1,5 +1,7 @@
 module Control.Joint.Schemes.UT (UT (..)) where
 
+import "comonad" Control.Comonad (Comonad (extract))
+import "comonad" Control.Comonad.Trans.Class (ComonadTrans (lower))
 import "transformers" Control.Monad.Trans.Class (MonadTrans (lift))
 
 import Control.Joint.Core (type (:.), type (:=))
@@ -13,3 +15,6 @@ instance Interpreted (UT t u) where
 
 instance (forall u . Functor u, Monad t) => MonadTrans (UT t) where
 	lift x = UT $ return <$> x
+
+instance Comonad t => ComonadTrans (UT t) where
+	lower (UT x) = extract <$> x